Add linting and switch AsErrChecker to a generic
This adds linting with golangci-lint; all linting issues were fixed as
part of this. Additionally, to better handle using `errors.As` in the
Check function of the AsErrChecker, it was switched to be a generic type
instead of trying to embed and properly use an interface. This allows us
to add a type constraint to it such that it only accepts valid error
values, and prevents linter errors from trying to use `interface{}` or
something gross like `*error`.