Add printf style Assertf and Requiref
This adds printf style Assert and Require functions, and changes the
semantics of the non-printf style Assert and Require functions. This
simplifies the functions a bit and avoids ambiguity around passing a
string value as the first value to those functions works; now if you
want printf style formatting you use the Assertf or Requiref functions,
and if you want to output a string as-is you use the Assert or Require
functions.